시야 확인 페이지 수정
