+++ src/pages/Login.jsx
@@ -1,0 +1,8 @@
+
+function Login() {
+    return(
+        <></>
+    )
+}
+
+export default Login;
