+++ src/pages/SignUp.jsx
@@ -1,0 +1,10 @@
+
+function SignUp() {
+
+    return(
+        <>
+        </>
+    )
+}
+
+export default SignUp;
